AYS-278 | Refactor : Get error messages from enum class, Add null case to invalidPermissionIds Data provider , move getPermissionIdsFromLastCreatedRole method to RoleDataSource class
