+++ src/test/java/org/ays/auth/datasource/RoleDataSource.java
@@ -7,6 +7,8 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
 
 @UtilityClass
 public class RoleDataSource {
@@ -94,6 +96,31 @@ public class RoleDataSource {
 
         } catch (SQLException exception) {
             throw new RuntimeException(exception);
+        }
+    }
+
+    public static List<String> getPermissionIdsFromLastCreatedRole(String roleId) {
+
+        String query = "SELECT PERMISSION_ID FROM AYS_ROLE_PERMISSION_RELATION WHERE ROLE_ID = ?";
+
+        try (Connection connection = AysDataSource.createConnection();
+             PreparedStatement preparedStatement = connection.prepareStatement(query)) {
+
+            preparedStatement.setString(1, roleId);
+
+            try (ResultSet resultSet = preparedStatement.executeQuery()) {
+                List<String> permissions = new ArrayList<>();
+
+                while (resultSet.next()) {
+                    String permissionId = resultSet.getString("PERMISSION_ID");
+                    permissions.add(permissionId);
+                }
+
+                return permissions;
+            }
+
+        } catch (SQLException exception) {
+            throw new RuntimeException("Error fetching permission IDs for role: " + roleId, exception);
         }
     }
