+++ src/test/java/org/ays/auth/tests/RoleCreateTest.java
@@ -1,7 +1,6 @@
 package org.ays.auth.tests;
 
 import io.restassured.response.Response;
-import org.ays.auth.datasource.PermissionDataSource;
 import org.ays.auth.datasource.RoleDataSource;
 import org.ays.auth.endpoints.AuthEndpoints;
 import org.ays.auth.endpoints.RoleEndpoints;
@@ -25,9 +24,8 @@ public class RoleCreateTest {
 
     @BeforeMethod
     private void setUpToken() {
-
         LoginPayload loginPayload = LoginPayload.generateAsTestVolunteerFoundationAdmin();
-        accessToken = loginAndGetAccessToken(loginPayload);
+        this.accessToken = this.loginAndGetAccessToken(loginPayload);
     }
 
     @Test(groups = {"Smoke", "Regression"})
@@ -63,7 +61,6 @@ public class RoleCreateTest {
         Response response = RoleEndpoints.create(role, accessToken);
         response.then()
                 .spec(AysResponseSpecs.expectConflictResponseSpec())
-                .body("header", equalTo("ALREADY EXIST"))
                 .body("message", equalTo("role already exist! name:" + existingRoleName + ""));
     }
 
@@ -88,7 +85,7 @@ public class RoleCreateTest {
         Response response = RoleEndpoints.create(role, accessToken);
         response.then()
                 .spec(AysResponseSpecs.expectBadRequestResponseSpec())
-                .body("subErrors[0].message", equalTo("must not be blank"));
+                .body("subErrors[0].message", equalTo(AysErrorMessage.MUST_NOT_BE_BLANK.getMessage()));
     }
 
     @Test(groups = {"Regression"}, dataProvider = "invalidPermissionIds", dataProviderClass = AysDataProvider.class)
@@ -105,18 +102,6 @@ public class RoleCreateTest {
     }
 
     @Test(groups = {"Regression"})
-    public void createRoleWithNullPermissionId() {
-
-        RoleCreatePayload role = RoleCreatePayload.generate();
-        role.setPermissionIds(null);
-
-        Response response = RoleEndpoints.create(role, accessToken);
-        response.then()
-                .spec(AysResponseSpecs.expectBadRequestResponseSpec())
-                .body("subErrors[0].message", equalTo("must not be empty"));
-    }
-
-    @Test(groups = {"Regression"})
     public void createRoleWithDuplicatePermissionIds() {
 
         RoleCreatePayload role = RoleCreatePayload.generate();
@@ -125,7 +110,7 @@ public class RoleCreateTest {
         String roleId = RoleDataSource.findLastCreatedRoleIdByInstitutionId(AysConfigurationProperty
                 .TestVolunteerFoundation.ID);
 
-        List<String> permissionIds = PermissionDataSource.getPermissionIdsFromLastCreatedRole(roleId);
+        List<String> permissionIds = RoleDataSource.getPermissionIdsFromLastCreatedRole(roleId);
         List<String> expectedPermissionIds = List.of("17dd50f6-61fe-4a30-a136-d9b80649e7fe");
 
         Assert.assertEquals(permissionIds.size(), 1, "Permission IDs should be unique in the database.");
@@ -134,6 +119,7 @@ public class RoleCreateTest {
 
     @Test(groups = {"Regression"})
     public void createRoleByUserWithoutRoleCreatePermission() {
+
         LoginPayload login = LoginPayload.generateAsTestVolunteerFoundationUser();
         String token = loginAndGetAccessToken(login);
 
@@ -141,10 +127,7 @@ public class RoleCreateTest {
 
         Response response = RoleEndpoints.create(role, token);
         response.then()
-                .spec(AysResponseSpecs.expectUnauthorizedResponseSpec())
-                .body("header", equalTo("AUTH ERROR"))
-                .body("isSuccess", equalTo(false));
-
+                .spec(AysResponseSpecs.expectUnauthorizedResponseSpec());
     }
 
     private String loginAndGetAccessToken(LoginPayload loginPayload) {
